adds skeleton for weasyprint and notes on difficulty as well as log; adds sample outputs; moves pdf conversion to utils
